+++ appinfo/update.php
@@ -1,3 +1,12 @@
 <?php
+/**
+ * ownCloud - fbsync
+ *
+ * This file is licensed under the Affero General Public License version 3 or
+ * later. See the COPYING file.
+ *
+ * @author NOIJN <fremulon@protonmail.com>
+ * @copyright NOIJN 2015
+ */
 
 ?>